CPUTuneCore: Add support for MacOS Tahoe
The dead end of Intel Macs is here! As we all know, Tahoe is the last
supported version for Intel Macs; after that only security updates is
provided, and there will likely no kernel version bump in the future.

The Darwin kernel that Tahoe uses is Darwin kernel 25, add support for
it, and remove that getKernelVersion() hack we're already have.

Famous last words: Tung Tung Tung Tung Tahoe.

This reverts commit b260ab47045bee3bff573007613c5926d0d76c38.

Signed-off-by: Tuan Anh <tuan73176@gmail.com>